feat: fjord
Add placeholder for coordination of the fjord network upgrade.
Experimental and meant to centralize discussion of what goes into
the fjord network upgrade in a single place and ahead of time.

Also using an experiment directory structure where the network
upgrade name is a directory and includes all of the changes for
that particular upgrade. It would be good to refactor the rest of
the repo to follow this standard so it is easy to know exactly what
is changing between network upgrades. At the moment, the information
is spread out across many files.