specs: remove dead devnet doc (#12)
Deletes an outdated document in the specs for the devnet.
This repo should not contain information about the monorepo
or development practices as it will easily get out of sync.
This document is already out of sync as it mentions hardhat,
the devnet has evolved a lot since this time.