interop: remove native ether sends
Remove native ether sends from the interop protocol to enable interop
with chains that use a custom gas token. ether will need to be wrapped
into WETH for being able to send ether between domains.

This simplifies the implementations of the bridge predeploys and removes
an upgrade transaction that mints ether to the messenger predeploy.
A design pattern similar to scroll was employed where cross domain ether
liquidity sits locked in a contract. This ether technically "does not
exist" from the point of view of the protocol, similar to ether locked
up in the validator deposit contract. It removes a huge security
backdoor where somehow this ether can be unlocked when it shouldn't be,
reducing the scope of the audit.