Cleanup Span Batch Limits
This modifies how the span batch limits are specified. Previously it was using
a protocol constant which was denominated in bytes (MAX_RLP_BYTES_PER_CHANNEL)
to limit the total size, the number of blocks, the number of transactions per
block, and the total number of transactions.

This PR modifies the spec to split out MAX_RLP_BYTES_PER_CHANNEL and
MAX_SPAN_BATCH_ELEMENT_COUNT. This is not a consensus change because
the values are the same. In addition, this removes the size check from
the span batch validation. This is also not a consensus change until the
Fjord hard fork because MAX_RLP_BYTES_PER_CHANNEL is checked prior to
decoding the span batch. After Fjord, the behavior will change when
the MAX_RLP_BYTES_PER_CHANNEL constant is increased.