maint: move protocol upgrades down in sidebar
Protocol upgrades are useful but less important than the core
protocol specifications. PR moves the information about these
upgrades down to the bottom of the protocol section and renames
network upgrades to protocol upgrades to account for the fact that
these upgrades are now stack-wide and not for a single network.