feat: op network execution payload envelope decoding (#149)
<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide:
https://github.com/alloy-rs/core/blob/main/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and
building the
documentation.
-->

<!-- ** Please select "Allow edits from maintainers" in the PR Options
** -->

## Motivation

Ports the necessary types for decoding a snappy+ssz encoded
`ExecutionPayloadEnvelope` from the OP CL network.

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? In some cases there is not a problem and this
can be
thought of as being the motivation for your change.
-->



## PR Checklist

- [x] Added Tests
- [x] Added Documentation

---------

Co-authored-by: refcell <abigger87@gmail.com>