chore: revert maili deps (#399)
Closes https://github.com/alloy-rs/op-alloy/issues/398

this reverts a bunch of stuff and removes all maili dependencies.

this is definitely incomplete but is the first step towards restoring
op-alloy as a maili independent repo.

the dep graph looks now as follows:
alloy -> op-alloy -> maili

maili can have all types limited to protocol impls (although debatable
if this is actually beneficial, imo this just slows down development
atm).

next steps are invert deps on maili side and clean things up there.

this repo would benefit from the flz stuff, so I suggest we move this
crate in here and only keep the brotli stuff in maili